+++ src/main/java/com/example/DevNote/repository/QuizzRepository.java
@@ -3,5 +3,5 @@ package com.example.DevNote.repository;
 import com.example.DevNote.model.Quizz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuizzRepository extends JpaRepository<Quizz, Long> {
+public interface QuizzRepository extends JpaRepository<Quizz, Integer> {
 }
