+++ src/main/java/com/example/DevNote/repository/SessionRepository.java
@@ -3,5 +3,5 @@ package com.example.DevNote.repository;
 import com.example.DevNote.model.Session;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SessionRepository extends JpaRepository<Session, Long> {
+public interface SessionRepository extends JpaRepository<Session, Integer> {
 }
