+++ src/main/java/com/example/DevNote/repository/TechRepository.java
@@ -3,5 +3,5 @@ package com.example.DevNote.repository;
 import com.example.DevNote.model.Tech;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TechRepository extends JpaRepository<Tech, Long> {
+public interface TechRepository extends JpaRepository<Tech, Integer> {
 }
