+++ src/main/java/com/example/DevNote/repository/UsersRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsersRepository extends JpaRepository<Users, Long> {  //UsersRepository hérite de l'interface JpaRepository. Cette dernière prend deux paramètres génériques : le type de
-    // l'entité (Users) et le type de la clé primaire (Long). En héritant de JpaRepository, UsersRepository bénéficie
+public interface UsersRepository extends JpaRepository<Users, Integer> {  //UsersRepository hérite de l'interface JpaRepository. Cette dernière prend deux paramètres génériques : le type de
+    // l'entité (Users) et le type de la clé primaire (Integer). En héritant de JpaRepository, UsersRepository bénéficie
     // automatiquement de plusieurs méthodes utiles pour la manipulation des données (comme save(), delete(), findById(),
     // findAll(), etc.).
